mcp-monarch/requirements.txt
# MCP Server dependencies
mcp>=1.26.0
anthropic>=0.84.0

# Telegram Bot
python-telegram-bot>=22.6

# AWS for agent deployment
boto3>=1.42.0

# Utilities
aiohttp>=3.13.0
python-dotenv>=1.2.0
